+++ ldj-spring-renderer-example/src/main/resources/com/infosupport/ldoc/springexample/amqp/template.adoc
