+++ ldj-spring-renderer-example/src/test/java/com/infosupport/ldoc/springexample/amqp/QueueInteractionTest.java
@@ -1,4 +1,4 @@
-package com.infosupport.ldoc.springexample.ampq;
+package com.infosupport.ldoc.springexample.amqp;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertIterableEquals;
